Refactor exception handling inspection
