+++ plugin/src/main/resources/META-INF/plugin.xml
@@ -103,17 +103,9 @@
                          groupName="PyNose"
                          enabledByDefault="true"
                          level="WARNING"
-                         implementationClass="org.jetbrains.research.pynose.plugin.inspections.unittest.ExceptionHandlingTestSmellUnittestInspection"
-                         displayName="Exception handling test unittest"
-                         shortName="ExceptionHandlingTestSmellUnittestInspection"/>
-        <localInspection language="Python"
-                         groupPath="Python"
-                         groupName="PyNose"
-                         enabledByDefault="true"
-                         level="WARNING"
-                         implementationClass="org.jetbrains.research.pynose.plugin.inspections.pytest.ExceptionHandlingTestSmellPytestInspection"
-                         displayName="Exception handling test pytest"
-                         shortName="ExceptionHandlingTestSmellPytestInspection"/>
+                         implementationClass="org.jetbrains.research.pynose.plugin.inspections.universal.ExceptionHandlingTestSmellInspection"
+                         displayName="Exception handling test"
+                         shortName="ExceptionHandlingTestSmellInspection"/>
         <localInspection language="Python"
                          groupPath="Python"
                          groupName="PyNose"
