+++ plugin/src/main/resources/inspectionDescriptions/ExceptionHandlingTestSmellInspection.html
