--- plugin/src/main/resources/inspectionDescriptions/ExceptionHandlingTestSmellPytestInspection.html
@@ -1,10 +1,0 @@
-<html>
-<body>
-
-<b>Exception Handling (pytest)</b> occurs when passing or failing of a test case
-is dependent on the production method explicitly throwing an exception.
-Instead, developers should utilize special functionality of testing frameworks for that,
-such as a pytest.raises() function.
-
-</body>
-</html>
